testing wallet withdrawals

+++ core/src/main/java/com/newtally/core/service/MerchantService.java
@@ -683,7 +683,7 @@ public class MerchantService extends AbstractService implements IAuthenticator {
 		return query.getResultList().get(0).toString();
 	}
 	
-	public void createWithdrawTransaction(Integer currencyId, String walletAddress, Double transactionAmount, Double commissionAmount, String commissionWalletAddress) {
+	public void createWithdrawTransaction(Integer currencyId, String walletAddress, long transactionAmount, long commissionAmount, String commissionWalletAddress) {
 	    EntityTransaction trn = em.getTransaction();
         trn.begin();
         try {
@@ -709,7 +709,7 @@ public class MerchantService extends AbstractService implements IAuthenticator {
         }
 	}
 	
-	public void createCommissionTransaction(Integer currencyId, String walletAddress, Double commissionAmount, Long withdrawlTransactionHistoryId) {
+	public void createCommissionTransaction(Integer currencyId, String walletAddress, long commissionAmount, Long withdrawlTransactionHistoryId) {
 		try {
             Query query = em.createNativeQuery("INSERT INTO commission_transaction_history(id, currency_id, to_wallet_address,"
                     + " merchant_id, transaction_date, withdrawl_transaction_history_id, commission_amount, transaction_status) "
